requirements.txt
flask
flask-cors
rembg
pillow
opencv-python
numpy
